requirements.txt
PyJWT==2.10.1
python-decouple==3.8
fastapi==0.115.11
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic-core==2.27.2
pydantic[email]==2.10.6
python-multipart==0.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
Jinja2==3.1.6
asyncpg==0.28.0
uvicorn==0.34.0
httpx==0.28.1
alembic==1.15.1
psycopg2-binary==2.9.9
redis==5.0.1
SQLAlchemy==2.0.39
bcrypt==4.0.1
email-validator==2.2.0
httpcore==1.0.7
slowapi==0.1.9